anthology_updater/requirements.txt
# requirements.txt
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0